Fixed a few issues that Niek pointed out.
I really should recode the AsyncStreamReader class to better handle situations where the reader is closed between reads. Ideally, I should print the rest of the stream before anything else. I'll have to take a deeper look into this. 

Also, I need to add events to LispPipe that notify a subscriber when a process is created, closed, or terminated unexpectedly.

git-svn-id: https://lispidedotnet.svn.codeplex.com/svn@13205 a83551a4-30f6-4d81-a974-c6ced450ddbf